DPAD-996: adding caption toggle button
